requirements.txt
arrow==0.15.2
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-cors-headers==3.1.1
django-countries==5.5
django-solo==1.1.3
djangorestframework==3.10.3
idna==2.8
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
